+++ package.json
@@ -3,8 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@rpch/sdk": "C:/dev/rpch/RPCh/packages/sdk",
-    "@tanstack/react-query": "^5.35.1",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -12,7 +10,6 @@
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "viem": "2.x",
-    "wagmi": "^2.8.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
