+++ src/App.css
@@ -2,37 +2,93 @@
   text-align: center;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+#whole {
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
 }
 
-.App-link {
-  color: #61dafb;
+#status {
+  display: none;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+#blocks-and-rest {
+  display: flex;
+  flex-direction: row;
+}
+
+#blocks {
+  max-height: calc( 100vh - 200px );
+  width: 150px;
+}
+
+#blocks-title {
+  margin-bottom: 8px;
+  text-align: center;
+  width: 150px;
+  margin-right: 16px;
+  margin-left: 16px;
+}
+
+#blocks-list {
+  max-height: calc( 100vh - 200px );
+  width: 150px;
+  overflow: auto;
+  margin-right: 16px;
+  margin-left: 16px;
+}
+
+.block-list-item {
+  cursor: pointer;
+  text-decoration: underline;
+  margin-bottom: 4px;
+}
+
+.weight400 {
+  font-weight: 400;
+}
+
+.weight700 {
+  font-weight: 700;
+}
+
+#explorer {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1
+}
+
+#transactions {
+  height: 330px;
+  margin-bottom: 16px;
+}
+
+#transactions-title {
+  margin-bottom: 8px;
+}
+
+#transactions-list {
+  max-height: 300px;
+  overflow: auto;
+  max-width: 660px;
+  width: 100%;
+  margin: auto;
+}
+
+.transactions-list-item  {
+  cursor: pointer;
+  text-decoration: underline;
+  margin-bottom: 4px;
+}
+
+#logs-title {
+  margin-bottom: 8px;
+}
+
+
+#json {
+  max-width: 100vw;
+  overflow-wrap: anywhere;
+  word-break: break-all;
+  white-space: pre-line;
+  margin: 16px;
 }
